refactor: add types.ts for spliting type code
